Release: April 1st 2026.

+++ L2J_Mobius_CT_2.6_HighFive/java/org/l2jmobius/gameserver/network/serverpackets/AbstractItemPacket.java
@@ -17,10 +17,10 @@
 package org.l2jmobius.gameserver.network.serverpackets;
 
 import org.l2jmobius.commons.network.WritableBuffer;
-import org.l2jmobius.gameserver.model.ItemInfo;
-import org.l2jmobius.gameserver.model.TradeItem;
+import org.l2jmobius.gameserver.model.item.holders.ItemInfo;
 import org.l2jmobius.gameserver.model.item.instance.Item;
 import org.l2jmobius.gameserver.model.itemcontainer.PlayerInventory;
+import org.l2jmobius.gameserver.network.holders.TradeItem;
 
 /**
  * @author UnAfraid
@@ -43,9 +43,9 @@ public abstract class AbstractItemPacket extends ServerPacket
 		buffer.writeInt(item.getObjectId()); // ObjectId
 		buffer.writeInt(item.getItem().getDisplayId()); // ItemId
 		buffer.writeLong(item.getCount()); // Quantity
-		buffer.writeByte(item.getItem().getType2()); // Item Type 2 : 00-weapon, 01-shield/armor, 02-ring/earring/necklace, 03-questitem, 04-adena, 05-item
-		buffer.writeByte(item.getCustomType1()); // Filler (always 0)
-		buffer.writeLong(item.getItem().getBodyPart().getMask()); // Slot : 0006-lr.ear, 0008-neck, 0030-lr.finger, 0040-head, 0100-l.hand, 0200-gloves, 0400-chest, 0800-pants, 1000-feet, 4000-r.hand, 8000-r.hand
+		buffer.writeShort(item.getItem().getType2()); // Item Type 2 : 00-weapon, 01-shield/armor, 02-ring/earring/necklace, 03-questitem, 04-adena, 05-item
+		buffer.writeShort(item.getCustomType1()); // Filler (always 0)
+		buffer.writeInt(item.getItem().getBodyPart().getMask()); // Slot : 0006-lr.ear, 0008-neck, 0030-lr.finger, 0040-head, 0100-l.hand, 0200-gloves, 0400-chest, 0800-pants, 1000-feet, 4000-r.hand, 8000-r.hand
 		buffer.writeShort(item.getEnchant()); // Enchant level (pet level shown in control item)
 		buffer.writeShort(0); // Equipped : 00-No, 01-yes
 		buffer.writeShort(item.getCustomType2());
